+++ macros/stages/snowflake/snowflake_create_stages_statement.sql
@@ -1,8 +1,0 @@
-{%- macro snowflake_create_or_replace_stage_statement(relation, sql) -%}
-
-    {{ log("Creating stages " ~ relation) }}
-CREATE OR REPLACE STAGE {{ relation.include(database=(not temporary), schema=(not temporary)) }}
-    {{ sql }}
-    ;
-
-{%- endmacro -%}